todo-7-module: added tasks
